gui makes call to change instrument
